resolviendo la actualización de stock

+++ src/components/Form/Form.jsx
@@ -1,4 +1,13 @@
-import { addDoc, collection, serverTimestamp } from 'firebase/firestore';
+import {
+    addDoc,
+    collection,
+    documentId,
+    getDocs,
+    query,
+    serverTimestamp,
+    where,
+    writeBatch,
+} from 'firebase/firestore';
 import { useContext, useState } from 'react';
 import { CartContext } from '../../context/CartContext';
 import { db } from '../../services/firebaseConfig';
@@ -14,28 +23,67 @@ const Form = () => {
 
     const totalPrice = total();
 
-    const handleSubmit = (e) => {
+    const handleSubmit = async (e) => {
         setLoading(true);
         e.preventDefault();
         //console.log({ name, lastName });
-        const order = {
-            buyer: { name, lastName },
-            items: cart,
-            total: totalPrice,
-            date: serverTimestamp(),
-        };
+        try {
+            const order = {
+                buyer: { name, lastName },
+                items: cart,
+                total: totalPrice,
+                date: serverTimestamp(),
+            };
 
-        const ordersCollection = collection(db, 'orders');
-        addDoc(ordersCollection, order)
-            .then((res) => {
-                //console.log(res.id);
-                setOrderId(res.id);
+            const ids = cart.map((prod) => prod.id);
+            //['asd656vhd','vdsjsjkl43j534','ds78cdc']
+            console.log(ids);
+
+            const productsRef = collection(db, 'productos');
+
+            const productsAddedFromFirestore = await getDocs(
+                query(productsRef, where(documentId(), 'in', ids))
+            );
+
+            const { docs } = productsAddedFromFirestore;
+
+            const outOfStock = [];
+
+            const batch = writeBatch(db);
+
+            docs.forEach((doc) => {
+                const dataDoc = doc.data();
+                const stockDB = dataDoc.stock;
+
+                const productAddedToCart = cart.find(
+                    (prod) => prod.id === doc.id
+                );
+
+                const prodQuantity = productAddedToCart?.cantidad;
+
+                if (stockDB >= prodQuantity) {
+                    batch.update(doc.ref, { stock: stockDB - prodQuantity });
+                } else {
+                    outOfStock.push({ id: doc.id, ...dataDoc });
+                }
+            });
+
+            if (outOfStock.length === 0) {
+                batch.commit();
+
+                const orderRef = collection(db, 'orders');
+                const orderAdded = await addDoc(orderRef, order);
+                setOrderId(orderAdded.id);
                 deleteAll();
-            })
-            .catch((error) => {
-                console.log(error);
-            })
-            .finally(() => setLoading(false));
+            } else {
+                console.log('No hay stock de algún producto');
+            }
+            console.log(outOfStock);
+        } catch (error) {
+            console.log(error);
+        } finally {
+            setLoading(false);
+        }
     };
 
     const handleChangeName = (e) => {
@@ -45,8 +93,6 @@ const Form = () => {
     const handleChangeLastName = (e) => {
         setLastName(e.target.value);
     };
-
-    console.log(orderId);
 
     if (orderId) {
         return (
